Use parallel routes to fetch repositories and add loading route
