+++ app/coding/@repos/loading.tsx
@@ -1,0 +1,8 @@
+export default function LoadingCoding() {
+  return (
+    <ul>
+      <li>Loading repositories from GitHub...</li>
+      <li>They should appear here soon.</li>
+    </ul>
+  );
+}
